make a jupyter server pytest plugin
