+++ setup.py
@@ -100,6 +100,9 @@ for more information.
         'console_scripts': [
             'jupyter-server = jupyter_server.serverapp:main',
             'jupyter-bundlerextension = jupyter_server.bundler.bundlerextensions:main',
+        ],
+        'pytest11': [
+            'pytest_jupyter_server = jupyter_server.pytest_plugin'
         ]
     },
 )
